Правки в пакетах моделей и разбиение serializers.py на аналогичные моделям пакеты
