+++ backend/api/models/recipe_tags.py
@@ -1,11 +1,11 @@
 from django.db import models
 
-from api.models.abstract_models import CookbookModel
+from api.models.abstract_models import CookbookBaseModel
 from api.models.recipe import Recipe
 from api.models.tag import Tag
 
 
-class RecipeTags(CookbookModel):
+class RecipeTags(CookbookBaseModel):
     recipe = models.ForeignKey(
         to=Recipe, verbose_name='Рецепт', on_delete=models.CASCADE
     )
@@ -13,7 +13,7 @@ class RecipeTags(CookbookModel):
         to=Tag, verbose_name='Тег', on_delete=models.CASCADE
     )
 
-    class Meta(CookbookModel.Meta):
+    class Meta(CookbookBaseModel.Meta):
         default_related_name = 'recipe_tags'
         verbose_name = 'Тег рецепта'
         verbose_name_plural = 'Теги рецептов'
