+++ backend/api/models/tag.py
@@ -1,10 +1,10 @@
 from django.db import models
 
-from api.models.abstract_models import CookbookModel
+from api.models.abstract_models import CookbookBaseModel
 from core.constants import LENGTH_CHARFIELD_32
 
 
-class Tag(CookbookModel):
+class Tag(CookbookBaseModel):
     # TODO: Написать manager и management-загрузку
 
     name = models.CharField(
@@ -18,7 +18,7 @@ class Tag(CookbookModel):
         unique=True
     )
 
-    class Meta(CookbookModel.Meta):
+    class Meta(CookbookBaseModel.Meta):
         verbose_name = 'Тег'
         verbose_name_plural = 'Теги'
         ordering = ['name']
