+++ backend/api/serializers/ingredient.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+
+from api.models import Ingredient
+
+
+class IngredientSerializer(serializers.ModelSerializer):
+    """Сериалайзер для ингредиентов."""
+
+    class Meta:
+        model = Ingredient
+        fields = ('id', 'name', 'measurement_unit',)
