+++ backend/api/serializers/recipe_ingredients.py
@@ -1,0 +1,29 @@
+from rest_framework import serializers
+
+from api.models import Ingredient, RecipeIngredients
+
+
+class RecipeIngredientsSetSerializer(serializers.ModelSerializer):
+    """Сериалайзер связующей рецепты+ингредиенты на запись."""
+
+    id = serializers.PrimaryKeyRelatedField(
+        queryset=Ingredient.objects.all()
+    )
+
+    class Meta:
+        model = RecipeIngredients
+        fields = ('id', 'amount')
+
+
+class RecipeIngredientsGetSerializer(serializers.ModelSerializer):
+    """Сериалайзер связующей рецепты+ингредиенты на чтение."""
+
+    id = serializers.ReadOnlyField(source='ingredient.id')
+    name = serializers.ReadOnlyField(source='ingredient.name')
+    measurement_unit = serializers.ReadOnlyField(
+        source='ingredient.measurement_unit'
+    )
+
+    class Meta:
+        model = RecipeIngredients
+        fields = ('id', 'name', 'measurement_unit', 'amount')
