+++ backend/api/serializers/tag.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+
+from api.models import Tag
+
+
+class TagSerializer(serializers.ModelSerializer):
+    """Сериалайзер для тегов."""
+
+    class Meta:
+        model = Tag
+        fields = ('id', 'name', 'slug',)
