+++ backend/core/constants.py
@@ -17,6 +17,16 @@ MIN_INGREDIENT_AMOUNT_ERROR = (
 )
 USER_EMAIL_ERROR = 'Данный электронный адрес уже используется.'
 USER_USERNAME_ERROR = 'Пользователь с таким ником уже существует.'
+SUPERUSER_STAFF_ERROR = 'Суперпользователь должен иметь is_staff=True.'
+
+# Help-текст для полей модели
+HELP_USER_IS_STAFF = (
+    'Определяет, может ли пользователь войти на страницу администратора.'
+)
+HELP_USER_IS_ACTIVE = (
+    'Определяет, следует ли считать пользователя активным. '
+    'Уберите флажок вместо удаления учетной записи.'
+)
 
 # Названия "схем" - префикс перед названием таблицы
 COOKBOOK = 'cookbook'  # Префикс моделей приложения api
