+++ backend/users/models/abstract_models.py
@@ -1,15 +1,9 @@
-from django.db import models
-
 from core.constants import AUTH
-from core.utils import to_snake_case
+from core.models import PrefixedDBModel
 
 
-class AuthBaseModel(models.Model):
+class AuthBaseModel(PrefixedDBModel):
+    prefix_name = AUTH
 
-    class Meta:
+    class Meta(PrefixedDBModel.Meta):
         abstract = True
-
-    def __init_subclass__(cls, **kwargs) -> None:
-        super().__init_subclass__(**kwargs)
-        table_name = to_snake_case(cls.__name__)
-        cls.Meta.db_table = f'{AUTH}_{table_name}'
