+++ backend/users/serializers/__init__.py
@@ -1,0 +1,17 @@
+from users.serializers.user import (
+    AvatarSerializer,
+    CurrentUserSerializer,
+    UserSerializer
+)
+from users.serializers.subscription import (
+    SubscriptionChangedSerializer,
+    SubscriptionGetSerializer
+)
+
+__all__ = [
+    'AvatarSerializer',
+    'CurrentUserSerializer',
+    'SubscriptionChangedSerializer',
+    'SubscriptionGetSerializer',
+    'UserSerializer'
+]
